New extension class added
